requirements.txt
llama-index
farm-haystack
fastapi
uvicorn
requests
python-dotenv
